+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The complete changelog for the Costs to Expect REST API, follows the format defi
 ## [v1.22.0] - 2019-08-xx
 ### Added
 - We have added PATCH support for categories; if authenticated, it is possible to update the selected category.
+- We have added PATCH support for subcategories; if authenticated, it is possible to update the selected subcategory.
 - We have added a Request utility class with helper methods for POST and PATCH request validation.
 
 ### Changed